Split BrokerReduceService code into ResultSetters
