+++ pinot-core/src/main/java/org/apache/pinot/core/query/reduce/resultsetter/AggregationResultSetter.java
@@ -1,0 +1,110 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.pinot.core.query.reduce.resultsetter;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import org.apache.pinot.common.metrics.BrokerMetrics;
+import org.apache.pinot.common.request.AggregationInfo;
+import org.apache.pinot.common.request.BrokerRequest;
+import org.apache.pinot.common.response.broker.AggregationResult;
+import org.apache.pinot.common.response.broker.BrokerResponseNative;
+import org.apache.pinot.common.utils.DataSchema;
+import org.apache.pinot.common.utils.DataTable;
+import org.apache.pinot.core.query.aggregation.function.AggregationFunction;
+import org.apache.pinot.core.query.aggregation.function.AggregationFunctionUtils;
+import org.apache.pinot.core.transport.ServerRoutingInstance;
+import org.apache.pinot.core.util.QueryOptions;
+
+
+/**
+ * Helper class to set Aggregation results into the BrokerResponseNative
+ */
+public class AggregationResultSetter implements ResultSetter {
+
+  private final AggregationFunction[] _aggregationFunctions;
+  private final boolean _preserveType;
+
+  AggregationResultSetter(BrokerRequest brokerRequest, AggregationFunction[] aggregationFunctions,
+      QueryOptions queryOptions) {
+    _aggregationFunctions = aggregationFunctions;
+    _preserveType = queryOptions.isPreserveType();
+  }
+
+  /**
+   * Sets aggregations results into BrokerResponseNative::AggregationResults
+   */
+  @Override
+  public void setResults(String tableName, DataSchema dataSchema, Map<ServerRoutingInstance, DataTable> dataTableMap,
+      BrokerResponseNative brokerResponseNative, BrokerMetrics brokerMetrics) {
+    if (dataTableMap.isEmpty()) {
+      return;
+    }
+
+    assert dataSchema != null;
+
+    int numAggregationFunctions = _aggregationFunctions.length;
+    Collection<DataTable> dataTables = dataTableMap.values();
+
+    // Merge results from all data tables.
+    Object[] intermediateResults = new Object[numAggregationFunctions];
+    for (DataTable dataTable : dataTables) {
+      for (int i = 0; i < numAggregationFunctions; i++) {
+        Object intermediateResultToMerge;
+        DataSchema.ColumnDataType columnDataType = dataSchema.getColumnDataType(i);
+        switch (columnDataType) {
+          case LONG:
+            intermediateResultToMerge = dataTable.getLong(0, i);
+            break;
+          case DOUBLE:
+            intermediateResultToMerge = dataTable.getDouble(0, i);
+            break;
+          case OBJECT:
+            intermediateResultToMerge = dataTable.getObject(0, i);
+            break;
+          default:
+            throw new IllegalStateException("Illegal column data type in aggregation results: " + columnDataType);
+        }
+        Object mergedIntermediateResult = intermediateResults[i];
+        if (mergedIntermediateResult == null) {
+          intermediateResults[i] = intermediateResultToMerge;
+        } else {
+          intermediateResults[i] = _aggregationFunctions[i].merge(mergedIntermediateResult, intermediateResultToMerge);
+        }
+      }
+    }
+
+    // Extract final results and set them into the broker response.
+    List<AggregationResult> reducedAggregationResults = new ArrayList<>(numAggregationFunctions);
+    for (int i = 0; i < numAggregationFunctions; i++) {
+      Serializable resultValue = AggregationFunctionUtils
+          .getSerializableValue(_aggregationFunctions[i].extractFinalResult(intermediateResults[i]));
+
+      // Format the value into string if required
+      if (!_preserveType) {
+        resultValue = AggregationFunctionUtils.formatValue(resultValue);
+      }
+      reducedAggregationResults.add(new AggregationResult(dataSchema.getColumnName(i), resultValue));
+    }
+    brokerResponseNative.setAggregationResults(reducedAggregationResults);
+  }
+}
